Refactor art service
